Initial commit
In this commit, I attempt to accomplish the tasks of dates 2016/05/20
and 2016/09/13. These tasks ask that I move the framework from the
project, with which it was tightly infused, to its own repository.
In addition, I must decouple the directory layout from the
framework components.

I've moved the framework to its own framework.