Change project status
My intention here is to develop the framework under a real-world
project. That's what I was doing initially. But I stopped in favour
to develop and launch the framework in public. It later dawned upon
me that developing the framework across in public and still needing
it for projects is not ideal, at all.

Why? Well, the undeveloped project will be in constant change as
its design (behaviour, APIs, etc.) are not final. And if I have
to be moving files and changing code across the framework in the
real-world project and the framework under development, then that'll
be very annoying and inefficient.

Therefore I've decided to put the public development of Geometry
on a hold until it has been matured.